Dump uncommitted changes from long ago :^)

+++ src/photograpy/filters/interpolate.py
@@ -23,4 +23,4 @@ class ReshapeFilter(InterpolationFilter):
 
         hh, ww = np.meshgrid(new_h_axis, new_w_axis, indexing='ij')
 
-        self.content = self.get_interpolator()((hh, ww)).astype(int)
+        self.content = self.get_interpolator()((hh, ww))
